requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.5
h11==0.14.0
idna==3.10
image==1.5.33
outcome==1.3.0.post0
pillow==11.1.0
PySocks==1.7.1
requests==2.32.3
selenium==4.28.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
